用户端分类，菜品代码，套餐代码的开发

+++ sky-server/src/main/java/com/sky/service/DishService.java
@@ -22,5 +22,5 @@ public interface DishService {
 
     List<Dish> selectAllDishByCategoryId(Integer categoryId);
 
-    List<DishVO> selectDishWithFlavorByCategoryId(Integer categoryId);
+    List<DishVO> selectDishWithFlavorByCategoryId(Long categoryId);
 }
